eagle_vision/kafka_consumers/requirements.txt
kafka-python>=2.0.2
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0
python-dotenv>=1.0.0
